Accept mix casing exponential notation
Exponential notation in JavaScript accept `e` and `E` for the exponential part of the number. Since you're relying on `from-exponential` which also accepts any casing for the syntax of a number written in exponential notation, I believe it's correct to accept this too, when converting a string into a BigNumber.